% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{buh.clust}
\alias{buh.clust}
\title{Perform Hierarchical Clustering on Asset Correlations}
\usage{
buh.clust(x)
}
\arguments{
\item{x}{A numeric matrix of asset returns.}
}
\value{
A list of asset clusters.
}
\description{
This function performs hierarchical clustering on asset correlations and returns the clustered groups.
}
