% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{insert.at}
\alias{insert.at}
\title{Insert Values at Specified Positions in a Vector}
\usage{
insert.at(a, pos, ...)
}
\arguments{
\item{a}{A vector.}

\item{pos}{A numeric vector specifying the positions to insert the values.}

\item{...}{Values to be inserted at the specified positions.}
}
\value{
A vector with the inserted values.
}
\description{
This function inserts specified values at given positions in a vector.
}
