% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.avg}
\alias{po.avg}
\title{Perform LASSO or Ridge Regression for Portfolio Optimization}
\usage{
po.avg(y0, x0, method = "LASSO")
}
\arguments{
\item{y0}{A numeric vector of response values.}

\item{x0}{A numeric matrix of predictors.}

\item{method}{The regularization method: "LASSO", "RIDGE", or "EN" (Elastic Net).}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function performs LASSO, Ridge, or Elastic Net regression for portfolio optimization.
}
