% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.bhu}
\alias{po.bhu}
\title{Perform Portfolio Optimization Using Clusters and LASSO}
\usage{
po.bhu(y0, x0, group, rep)
}
\arguments{
\item{y0}{A numeric vector of response values.}

\item{x0}{A numeric matrix of predictors.}

\item{group}{A list of asset clusters.}

\item{rep}{The number of repetitions for optimization.}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function performs portfolio optimization using clustering and LASSO regularization.
}
