% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.cols}
\alias{po.cols}
\title{Perform Simple Linear Model for Portfolio Optimization}
\usage{
po.cols(y0, x0)
}
\arguments{
\item{y0}{A numeric vector of response values.}

\item{x0}{A numeric matrix of predictors.}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function uses simple linear regression to perform portfolio optimization.
}
