% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.grossExp}
\alias{po.grossExp}
\title{Perform Gross Exposure Portfolio Optimization}
\usage{
po.grossExp(y0, x0, method = "NOSHORT")
}
\arguments{
\item{y0}{A numeric vector of response values.}

\item{x0}{A numeric matrix of predictors.}

\item{method}{The regularization method: "NOSHORT" or "EQUAL".}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function performs gross exposure portfolio optimization using LASSO.
}
