% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.JM}
\alias{po.JM}
\title{Perform James-Stein Portfolio Optimization}
\usage{
po.JM(x0)
}
\arguments{
\item{x0}{A numeric matrix of asset returns.}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function uses James-Stein estimation for portfolio optimization.
}
