% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.SW.lasso}
\alias{po.SW.lasso}
\title{Perform LASSO Regularization with Stochastic Weight Portfolio Optimization}
\usage{
po.SW.lasso(y0, x0, b, sample)
}
\arguments{
\item{y0}{A numeric vector of response values.}

\item{x0}{A numeric matrix of predictors.}

\item{b}{Number of assets to select in each sample.}

\item{sample}{Number of random samples to generate.}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function performs portfolio optimization using LASSO regularization and stochastic weight selection.
}
