% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_fun_test.R
\name{po.TZT}
\alias{po.TZT}
\title{Perform Portfolio Optimization Using TZT Method}
\usage{
po.TZT(x0, gamma)
}
\arguments{
\item{x0}{A numeric matrix of asset returns.}

\item{gamma}{A numeric parameter for the TZT method.}
}
\value{
A numeric vector of optimized portfolio weights.
}
\description{
This function performs portfolio optimization using the TZT method.
}
