% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Portfolio Analysis}
\usage{
prepare_data(
  dat,
  date_column_index = 1,
  start_date = "19990101",
  end_date = "20231231"
)
}
\arguments{
\item{dat}{A data frame or matrix where the first column is the date and the remaining columns are the data.}

\item{date_column_index}{The index of the date column in the input data. Default is 1.}

\item{start_date}{A character string specifying the start date for filtering the data in 'YYYYMMDD' format.
Default is '19990101'.}

\item{end_date}{A character string specifying the end date for filtering the data in 'YYYYMMDD' format.
Default is '20231231'.}
}
\value{
A list containing the following components:
\describe{
  \item{x}{A matrix of the filtered data with missing values handled.}
  \item{mon}{A vector of integers representing the number of months from the first date in the data.}
  \item{count}{A vector of the number of entries per month.}
  \item{Date}{A vector of Date objects representing the filtered dates.}
}
}
\description{
This function prepares the input data by filtering based on a specified date range, removing the date column,
and handling missing values. It also generates time-related columns and returns the processed data.
}
\examples{
\dontrun{
data <- data.frame(Date = c("19990101", "19990102", "19990103"),
                   Var1 = c(1, 2, -99.99),
                   Var2 = c(3, 4, 5))
result <- prepare_data(data, date_column_index = 1)
print(result)}

}
